Fixed messages not being written to the database
The continue in the exception catching of Handle Message was causing the thread to remain "alive" but there was no activity on it, thus causing messages to not be written to the database after a short while. For some reason the error is not being written to the console, and I'm not sure why yet.

Also, the queue needs to be converted to a list of queues and each IRC Bot thread should get it's own queue, because right now more messages are put into the queue than this program can process.